réorganisation des plugins
