+++ program/Engine/System.js
@@ -2,7 +2,7 @@ const Player = require('./Player');
 const Mobile = require('./Mobile');
 const Area = require('./Area');
 const Emitter = require('events');
-const DataManager = require('../DataManager/index');
+const DataManager = require('../ResourceLoader/index');
 const logger = require('../Logger/index');
 const o876 = require('../o876/index');
 const Vector = o876.geometry.Vector;
