--- public/scripts/src/store/modules.js
@@ -1,13 +1,0 @@
-import chat from './chat';
-import clients from './clients';
-import gameMenu from './gameMenu';
-import player from './player';
-import ui from './ui';
-
-export default {
-    chat,
-    clients,
-    gameMenu,
-    player,
-    ui
-};
