+++ public/scripts/src/store/modules/chat/getters.js
