+++ public/scripts/src/store/modules/index.js
@@ -1,0 +1,13 @@
+import chat from './chat/index';
+import clients from './clients/index';
+import gameMenu from './gameMenu/index';
+import player from './player/index';
+import ui from './ui/index';
+
+export default {
+    chat,
+    clients,
+    gameMenu,
+    player,
+    ui
+};
