+++ public/scripts/src/store/modules/player/index.js
