+++ public/scripts/src/store/modules/player/inventory/index.js
