+++ public/scripts/src/store/modules/player/inventory/types.js
