+++ public/scripts/src/store/plugins/engine.js
@@ -1,4 +1,4 @@
-import STATUS from "../../../../program/consts/status";
+import STATUS from "../../../../../program/consts/status";
 
 const OVERLAY = true;
